+++ syndicate/core/handlers.py
@@ -129,10 +129,16 @@ def syndicate():
 @click.option('--errors_allowed', is_flag=True,
               help='Flag to return successful response even if some tests '
                    'fail')
+@click.option('--skip_tests', is_flag=True, default=False,
+              help='Flag to not run tests')
 @verbose_option
 @timeit(action_name=TEST_ACTION)
-def test(suite, test_folder_name, errors_allowed):
+def test(suite, test_folder_name, errors_allowed, skip_tests):
     """Discovers and runs tests inside python project configuration path."""
+    if skip_tests:
+        click.echo('Skipping tests...')
+        return
+
     click.echo('Running tests...')
     import subprocess
     from syndicate.core import CONFIG
@@ -175,12 +181,15 @@ def test(suite, test_folder_name, errors_allowed):
 @click.option('--force_upload', '-F', is_flag=True, default=False,
               help='Flag to override existing bundle with the same name')
 @click.option('--errors_allowed', is_flag=True, default=False,
-              help='Flag to continue bundle building if some tests fail')
+              help='Flag to continue building the bundle if any errors occur '
+                   'while building dependencies or tests fail')
+@click.option('--skip_tests', is_flag=True, default=False,
+              help='Flag to skip lambda tests')
 @verbose_option
 @click.pass_context
 @check_deploy_bucket_exists
 @timeit(action_name=BUILD_ACTION)
-def build(ctx, bundle_name, force_upload, errors_allowed):
+def build(ctx, bundle_name, force_upload, errors_allowed, skip_tests):
     """
     Builds bundle of an application
     """
@@ -189,8 +198,9 @@ def build(ctx, bundle_name, force_upload, errors_allowed):
                    f'in deploy bucket. Please use another bundle '
                    f'name or delete the bundle')
         return
-    ctx.invoke(test, errors_allowed=errors_allowed)
-    ctx.invoke(assemble, bundle_name=bundle_name, force_upload=force_upload)
+    ctx.invoke(test, errors_allowed=errors_allowed, skip_tests=skip_tests)
+    ctx.invoke(assemble, bundle_name=bundle_name,
+               errors_allowed=errors_allowed, force_upload=force_upload)
     ctx.invoke(package_meta, bundle_name=bundle_name)
     ctx.invoke(upload, bundle_name=bundle_name, force_upload=force_upload)
 
@@ -630,9 +640,12 @@ def profiler(bundle_name, deploy_name, from_date, to_date):
               help='Identifier that indicates whether a locally existing'
                    ' bundle should be deleted and a new one created using'
                    ' the same path.')
+@click.option('--skip_tests', is_flag=True, default=False,
+              help='Flag to not run tests')
 @verbose_option
 @timeit(action_name=ASSEMBLE_JAVA_MVN_ACTION)
-def assemble_java_mvn(bundle_name, project_path, force_upload):
+def assemble_java_mvn(bundle_name, project_path, force_upload, skip_tests,
+                      **kwargs):
     """
     Builds Java lambdas
 
@@ -640,14 +653,15 @@ def assemble_java_mvn(bundle_name, project_path, force_upload):
     :param bundle_name: name of the bundle
     :param project_path: path to project folder
     :param force_upload: force upload identification
+    :param skip_tests: force skipping tests
     :return:
     """
     click.echo(f'Command compile java project path: {project_path}')
     assemble_artifacts(bundle_name=bundle_name,
                        project_path=project_path,
                        runtime=RUNTIME_JAVA,
-                       force_upload=force_upload
-                       )
+                       force_upload=force_upload,
+                       skip_tests=skip_tests)
     click.echo('Java artifacts were prepared successfully.')
 
 
@@ -670,9 +684,12 @@ def assemble_java_mvn(bundle_name, project_path, force_upload):
               help='Identifier that indicates whether a locally existing'
                    ' bundle should be deleted and a new one created using'
                    ' the same path.')
+@click.option('--errors_allowed', is_flag=True, default=False,
+              help='Flag to continue building the bundle if any errors occur '
+                   'while building dependencies')
 @verbose_option
 @timeit(action_name=ASSEMBLE_PYTHON_ACTION)
-def assemble_python(bundle_name, project_path, force_upload):
+def assemble_python(bundle_name, project_path, force_upload, errors_allowed):
     """
     Builds Python lambdas
 
@@ -680,14 +697,15 @@ def assemble_python(bundle_name, project_path, force_upload):
     :param bundle_name: name of the bundle
     :param project_path: path to project folder
     :param force_upload: force upload identification
+    :param errors_allowed: allows to ignore dependency errors
     :return:
     """
     click.echo(f'Command assemble python: project_path: {project_path} ')
     assemble_artifacts(bundle_name=bundle_name,
                        project_path=project_path,
                        runtime=RUNTIME_PYTHON,
-                       force_upload=force_upload
-                       )
+                       force_upload=force_upload,
+                       errors_allowed=errors_allowed)
     click.echo('Python artifacts were prepared successfully.')
 
 
@@ -710,7 +728,7 @@ def assemble_python(bundle_name, project_path, force_upload):
                    ' the same path.')
 @verbose_option
 @timeit(action_name=ASSEMBLE_NODE_ACTION)
-def assemble_node(bundle_name, project_path, force_upload):
+def assemble_node(bundle_name, project_path, force_upload, **kwargs):
     """
     Builds NodeJS lambdas
 
@@ -748,7 +766,7 @@ def assemble_node(bundle_name, project_path, force_upload):
                    ' the same path.')
 @verbose_option
 @timeit(action_name=ASSEMBLE_DOTNET_ACTION)
-def assemble_dotnet(bundle_name, project_path, force_upload):
+def assemble_dotnet(bundle_name, project_path, force_upload, **kwargs):
     """
     Builds DotNet lambdas
 
@@ -816,10 +834,13 @@ RUNTIME_LANG_TO_BUILD_MAPPING = {
               help='Identifier that indicates whether a locally existing'
                    ' bundle should be deleted and a new one created using'
                    ' the same path.')
+@click.option('--errors_allowed', is_flag=True, default=False,
+              help='Flag to continue building the bundle if any errors occur '
+                   'while building dependencies. Only for Python runtime.')
 @verbose_option
 @click.pass_context
 @timeit(action_name=ASSEMBLE_ACTION)
-def assemble(ctx, bundle_name, force_upload):
+def assemble(ctx, bundle_name, force_upload, errors_allowed):
     """
     Builds the application artifacts
 
@@ -827,6 +848,8 @@ def assemble(ctx, bundle_name, force_upload):
     :param ctx:
     :param bundle_name: name of the bundle to which the artifacts
         will be associated
+    :param force_upload: force upload identification
+    :param errors_allowed: allows to ignore dependency errors. Only for python
     :return:
     """
     click.echo(f'Building artifacts, bundle: {bundle_name}')
@@ -840,7 +863,8 @@ def assemble(ctx, bundle_name, force_upload):
             func = RUNTIME_LANG_TO_BUILD_MAPPING.get(key)
             if func:
                 ctx.invoke(func, bundle_name=bundle_name,
-                           project_path=value, force_upload=force_upload)
+                           project_path=value, force_upload=force_upload,
+                           errors_allowed=errors_allowed)
             else:
                 click.echo(f'Build tool is not supported: {key}')
     else: